+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Changed
 - Implement modify scan config mutations [#2650](https://github.com/greenbone/gsa/pull/2650)
+- Implement schedule detailspage queries and mutations [#2659](https://github.com/greenbone/gsa/pull/2659)
 - Implement graphql queries and mutations on scanconfig listpage [#2621](https://github.com/greenbone/gsa/pull/2621)
 - Don't send scanner and config values in task dialog if task status is not new [#2598](https://github.com/greenbone/gsa/pull/2598)
 - Refactor task detailspage to not use withEntityContainer [#2596](https://github.com/greenbone/gsa/pull/2596)